+++ src/externs/omid-global.js
@@ -1,3 +1,5 @@
+/** @externs */
+
 /**
  * @type {!Window}
  */