+++ src/validation-verification-script/validation-verification-client.js
@@ -2,8 +2,9 @@ goog.module('omid.validationVerificationScript.ValidationVerificationClient');
 const {packageExport} = goog.require('omid.common.exporter');
 const {AdEventType} = goog.require('omid.common.constants');
 const VerificationClient = goog.require('omid.verificationClient.VerificationClient');
+const {isTopWindowAccessible, resolveGlobalContext} = goog.require('omid.common.windowUtils');
 /** @const {string} the default address for the logs.*/
-const DefaultLogServer = 'http://iabtechlab.com:66/sendMessage?msg=';
+const DefaultLogServer = 'http://localhost:66/sendmessage?msg=';
 
 /**
  * OMID ValidationVerificationClient.
@@ -46,6 +47,11 @@ class ValidationVerificationClient {
      * @param {number} timestamp of the event
      */
     logMessage_(message, timestamp) {
+        if (message.hasOwnProperty('type')) {
+            if (message['type'] === 'sessionStart') {
+                message.data.context['friendlyToTop'] = isTopWindowAccessible(resolveGlobalContext());
+            }
+        }
         const log = (new Date(timestamp)).toLocaleString()+ '::' + JSON.stringify(message);
         console.log(log);
         this.sendUrl_(log);